fix(editor) Removed a circular ref that was breaking the loaders test
